Refactor XML normalization shape conversion
